Change getFolderStats to commands.js
getFolderStatsをcommandsモジュールに委譲
